Admin create flight

+++ src/main/java/com/example/demo/Controllers/FlightController.java
@@ -36,16 +36,21 @@ public class FlightController {
     }
 
     @PostMapping("/addFlight")
-    public String createFlight(@RequestBody Flight newFlight){
+    public ResponseEntity<Void> createFlight(@RequestBody Flight newFlight){
         newFlight.setId(sequenceGeneratorService.getSequenceNumber(SEQUENCE_NAME));
         flightService.createFlight(newFlight);
-        return "Added flight with id: " + newFlight.getId();
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteFlight(@PathVariable Integer id){
-        flightService.deleteFlight(id);
-        return "Flight deleted ";
+    public ResponseEntity<Void> deleteFlight(@PathVariable Integer id){
+        Flight flight = flightService.findById(id).orElse(null);
+        if(flight == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else{
+            flightService.deleteFlight(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
     }
 
 }
